slime.el: fix inspector asynchronicity
* slime.el: slime-open-inspectoror: save-excursion here doesn't work
because the results come asynchronously and the point goes to the
bottom. What is needed is to restore the point in a
continuation which can run after the slime-inspector-insert-content
finishes its job. (slime-inspector-fetch-chunk): signature changed
to include  restore-point-cont argument which we ccall when the
chunks are inserted.

btw slime-insert-chunk-content is  probably an example of where fp is
not the most appropriate